Split parser from interval
