+++ src/jp_range/__init__.py
@@ -1,5 +1,6 @@
 """Utilities for parsing Japanese numeric ranges."""
 
-from .interval import Interval, parse_jp_range
+from .interval import Interval
+from .parser import parse_jp_range
 
 __all__ = ["Interval", "parse_jp_range"]
